feat: Update image-related components
Update the minor CSS on SectionHero and HomeHero components for the
better responsiveness. Remove unnecessary sectionHero.consts and move consts
into component. Add placeholder 'blur' to sectionContent's images.

feat: Update the minor css
