+++ app/_components/section/sectionContent/index.tsx
@@ -7,26 +7,29 @@ import { sectionContents } from '../section.consts';
 import { cx } from 'class-variance-authority';
 import { configsTransition } from '@/transition/transition.configs';
 import { SmoothTransitionWithDefaultConfigs } from '@/transition/smoothTransitionWithDefaultConfigs';
+import { TypesImageWithRemotePlaceholder } from '@/next/imageWithRemotePlaceholder/imageWithRemotePlaceholder.types';
 
 const styleImageWrapper = 'relative w-80 h-auto rounded-xl shadow-2xl shadow-blue-500/40 ring-purple-300/10';
 const styleImage = 'h-auto w-full rounded-xl drop-shadow-2xl';
 const styleImageFrame = 'md:h-[440px]';
-const optionsImageSpotlight = {
+const optionsImageSpotlight: TypesImageWithRemotePlaceholder = {
   width: 504,
   height: 689,
   className: styleImage,
   src: PATH_IMAGE['contentFocus'],
   sizes: '(max-width: 748px) 70vw, 33vw',
   alt: 'content focus image',
+  placeholder: 'blur',
   priority: true,
 };
-const optionsImageOverload = {
+const optionsImageOverload: TypesImageWithRemotePlaceholder = {
   width: 494,
   height: 650,
   className: styleImage,
   src: PATH_IMAGE['contentOrganize'],
   sizes: '(max-width: 748px) 70vw, 33vw',
   alt: 'content organize image',
+  placeholder: 'blur',
   priority: true,
 };
