+++ app/_components/section/sectionHero/index.tsx
@@ -3,15 +3,26 @@ import { SmoothTransition } from '@/transition/smoothTransition';
 import Link from 'next/link';
 import { SignInButton } from '@/button/signInButton';
 import { SmoothTransitionWithDivRef } from '@/transition/smoothTransitionWithDivRef';
-import { PATH_HOME } from '@/_lib/consts/assertion.consts';
+import { PATH_HOME, PATH_IMAGE } from '@/_lib/consts/assertion.consts';
 import { STYLE_BLUR_GRADIENT_R_LG } from '@/_lib/consts/style.consts';
 import { ImageWithRemotePlaceholder } from '@/_components/next/imageWithRemotePlaceholder';
 import { configsSignInButton } from '@/button/button.configs';
-import { optionsSectionHeroWithImage } from './sectionHero.consts';
 import { sectionContents } from '../section.consts';
 import { cx } from 'class-variance-authority';
 import { configsTransition } from '@/transition/transition.configs';
 import { SmoothTransitionWithDefaultConfigs } from '@/transition/smoothTransitionWithDefaultConfigs';
+import { TypesImageWithRemotePlaceholder } from '@/next/imageWithRemotePlaceholder/imageWithRemotePlaceholder.types';
+
+export const optionsSectionHeroWithImage: TypesImageWithRemotePlaceholder = {
+  width: 961,
+  height: 754,
+  className: 'h-auto rounded-2xl ring-2 ring-slate-300/20 drop-shadow-2xl will-change-transform',
+  src: PATH_IMAGE['demo'],
+  sizes: '90vw',
+  alt: 'demo application image',
+  placeholder: 'blur',
+  priority: true,
+};
 
 export const SectionHero = async () => {
   const divContainer_id = 'sectionHero';
